Rudimentary support for obserivng using the new iOS friendly watcher?
