+++ Fileaway.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -2,6 +2,15 @@
   "object": {
     "pins": [
       {
+        "package": "DirectoryWatcher",
+        "repositoryURL": "https://github.com/GianniCarlo/DirectoryWatcher.git",
+        "state": {
+          "branch": null,
+          "revision": "76af69f1d19146b0c5d3e38747bb5eabe547143b",
+          "version": "2.7.3"
+        }
+      },
+      {
         "package": "EonilFSEvents",
         "repositoryURL": "https://github.com/eonil/FSEvents",
         "state": {
